+++ vegitable/shops/views.py
@@ -16,40 +16,12 @@ from .models import ExpenditureEntry, PattiEntry, PattiEntryList, SalesBillEntry
     ArrivalGoods, CustomerLedger, FarmerLedger, CreditBillEntry, CreditBillHistory, Index
 import datetime
 from .report.report import Report
-from .utility import consolidate_result_for_report, get_float_number
+from .utility import consolidate_result_for_report, get_float_number, getDate_from_string
 from django.db.models import Sum, F, Q
 
 
 def index(request):
     return render(request, 'index.html')
-
-
-def getDate_from_string(stringDate: str):
-    mystringDate = str(stringDate).split("-")
-    return datetime.date(int(mystringDate[0]), int(mystringDate[1]), int(mystringDate[2]))
-
-
-@csrf_protect
-def get_authenticate(request):
-    user = auth.authenticate(
-        username=request.POST['username'], password=request.POST['password'])
-    if user is not None:
-        auth.login(request, user)
-        return home(request)
-    else:
-        messages.info(request, 'Invalid Email or Password')
-        return redirect('index')
-
-
-def home_next_page(request, page_number):
-    return home(request, current_page=page_number + 1)
-
-
-def home_prev_page(request, page_number):
-    if page_number > 1:
-        return home(request, current_page=page_number - 1)
-    else:
-        return home(request)
 
 
 def customer_ledger_next_page(request, page_number):
@@ -107,31 +79,7 @@ def sales_bill_prev_page(request, page_number):
         return sales_bill_entry(request)
 
 
-@csrf_protect
-def home(request, current_page=1):
-    if request.user.is_authenticated:
-        shop_detail_object = Shop.objects.get(shop_owner=request.user.id)
-        arrival_entry_detail = None
-        try:
 
-            arrival_entry_detail = ArrivalEntry.objects.filter(shop=shop_detail_object).filter(
-                Empty_data=False).order_by('-id')
-            items_per_page = 10
-            paginator = Paginator(arrival_entry_detail, items_per_page)
-            arrival_entry_detail = paginator.get_page(current_page)
-
-        except Exception as error:
-            print(error)
-
-        return render(request, 'Entry/Arrival/home.html',
-                      {
-                          'shop_details': shop_detail_object,
-                          'arrival_detail': arrival_entry_detail,
-                          'current_page': current_page,
-
-                      })
-
-    return render(request, 'index.html')
 
 
 @csrf_protect
@@ -349,27 +297,7 @@ def add_farmer_ledger(request):
     return render(request, 'index.html')
 
 
-def patti_list(request, current_page=1):
-    if request.user.is_authenticated:
-        shop_detail_object = Shop.objects.get(shop_owner=request.user.id)
-        patti_entry_detail = None
-        try:
-            patti_entry_detail = PattiEntry.objects.filter(shop=shop_detail_object).filter(Empty_data=False).order_by(
-                '-id')
-            items_per_page = 10
-            paginator = Paginator(patti_entry_detail, items_per_page)
-            patti_entry_detail = paginator.get_page(current_page)
-        except Exception as error:
-            print(error)
 
-        return render(request, 'Entry/Patti/patti.html',
-                      {
-                          'shop_details': shop_detail_object,
-                          'patti_entry_detail': patti_entry_detail,
-                          'current_page': current_page
-                      })
-
-    return render(request, 'index.html')
 
 
 def sales_bill_entry(request, current_page=1):
@@ -471,23 +399,7 @@ def logout(request):
     return render(request, 'index.html')
 
 
-def add_new_arrival_entry(request):
-    if request.user.is_authenticated:
-        today = date.today()
-        shop_detail_object = Shop.objects.get(shop_owner=request.user.id)
-        index = get_object_or_404(Index, shop=shop_detail_object)
 
-        arrival_index = {
-            'arrival_entry_prefix': index.arrival_entry_prefix,
-            'arrival_entry_counter': int(index.arrival_entry_counter)+1
-        }
-
-        return render(request, 'Entry/Arrival/modify_arrival_entry.html', {
-            "today": today,
-            "new": True,
-            "arrival_detail": arrival_index
-        })
-    return render(request, 'index.html')
 
 
 def add_new_expenditure_entry(request):
@@ -496,29 +408,7 @@ def add_new_expenditure_entry(request):
     return render(request, 'index.html')
 
 
-def add_new_patti_entry(request):
-    if request.user.is_authenticated:
-        today = date.today()
-        shop_detail_object = Shop.objects.get(shop_owner=request.user.id)
 
-        patti_bill_detail = PattiEntry(
-            lorry_no="",
-            date=today,
-            advance=0,
-            farmer_name="",
-            total_weight=0,
-            hamali=0,
-            net_amount=0,
-            shop=shop_detail_object,
-        )
-        patti_bill_detail.save()
-
-        return render(request, 'Entry/Patti/modify_patti_entry.html',
-                      {'patti_bill_detail': patti_bill_detail,
-                       "today": today,
-                       "NEW": True}
-                      )
-    return render(request, 'index.html')
 
 
 def navigate_to_add_sales_bill_entry(request):
@@ -742,19 +632,7 @@ def total_amount_expenditure_entry(request):
     return render(request, 'index.html')
 
 
-@csrf_protect
-def modify_arrival(request, arrival_id):
-    if request.user.is_authenticated:
-        arrival_entry_obj = ArrivalEntry.objects.get(pk=arrival_id)
-        arrival_goods_objs = ArrivalGoods.objects.filter(
-            arrival_entry=arrival_entry_obj).order_by('-id')
 
-        today = arrival_entry_obj.date
-
-        return render(request, 'Entry/Arrival/modify_arrival_entry.html',
-                      {'arrival_detail': arrival_entry_obj, 'arrival_goods_objs': arrival_goods_objs, 'new': False,
-                       "today": today})
-    return render(request, 'index.html')
 
 
 @api_view(('GET',))
@@ -818,123 +696,8 @@ def get_arrival_goods_list(request):
     return Response(data, status=status.HTTP_200_OK)
 
 
-@csrf_protect
-def add_arrival(request):
-    if request.user.is_authenticated:
-        shop_detail_object = Shop.objects.get(shop_owner=request.user.id)
-
-        if str(request.POST['new']) == "True":
-            arrival_Entry_Obj = ArrivalEntry(
-                arrival_id=request.POST['arrival_id'],  # for indexing
-                gp_no=request.POST['gp_number'],
-                date=getDate_from_string(request.POST['arrival_entry_date']),
-                patti_name=request.POST['patti_name'],
-                total_bags=request.POST['total_number_of_bags'],
-                lorry_no=request.POST['lorry_number'],
-                shop=shop_detail_object,
-                Empty_data=False)
-        else:
-            arrival_Entry_Obj = ArrivalEntry.objects.get(
-                id=request.POST['id'])  # object to update
-            arrival_Entry_Obj.gp_no = request.POST['gp_number']
-            arrival_Entry_Obj.date = getDate_from_string(
-                request.POST['arrival_entry_date'])
-            arrival_Entry_Obj.patti_name = request.POST['patti_name']
-            arrival_Entry_Obj.total_bags = request.POST['total_number_of_bags']
-            arrival_Entry_Obj.lorry_no = request.POST['lorry_number']
-            arrival_Entry_Obj.shop = shop_detail_object
-            arrival_Entry_Obj.Empty_data = False
-
-        arrival_Entry_Obj.save()
-        print(f"New arrival entry  = {arrival_Entry_Obj.id}")
-
-        if str(request.POST['new']) == "True":
-            index_obj = get_object_or_404(Index, shop=shop_detail_object)
-
-            # Increment the arrival_entry_counter
-            index_obj.arrival_entry_counter += 1
-            index_obj.save()
-
-        add_arrival_goods_item(request, list(
-            request.POST), arrival_Entry_Obj, shop_detail_object)
-
-        return home(request)
-
-    return render(request, 'index.html')
 
 
-def add_arrival_goods_item(request, request_list, arrival, shop_obj):
-    if request.user.is_authenticated:
-        former_name_list = []
-        item_name_list = []
-        qty_list = []
-        weight_list = []
-        remarks_list = []
-        arrival_goods_id = []
-        advance_amount_list = []
-
-        for i in request_list:
-            former_name_regrex = re.search("^.*_farmer_name$", i)
-            if former_name_regrex:
-                former_name_list.append(i)
-
-            item_name_regrex = re.search("^.*_item_name$", i)
-            if item_name_regrex:
-                item_name_list.append(i)
-
-            qty_regrex = re.search("^.*_qty$", i)
-            if qty_regrex:
-                qty_list.append(i)
-
-            weight_regrex = re.search("^.*_weight$", i)
-            if weight_regrex:
-                weight_list.append(i)
-
-            remark_regrex = re.search("^.*_remark$", i)
-            if remark_regrex:
-                remarks_list.append(i)
-
-            advance_amount_regrex = re.search("^.*_advance_amount$", i)
-            if advance_amount_regrex:
-                advance_amount_list.append(i)
-
-            arrival_goods_id_regrex = re.search("^.*_arrival_goods_id$", i)
-            if arrival_goods_id_regrex:
-                arrival_goods_id.append(i)
-
-        for i in range(0, len(former_name_list)):
-            if "modify" in item_name_list[i]:
-                local_id = request.POST[arrival_goods_id[i]].split("_")[0]
-                # object to update
-                arrival_Goods_obj = ArrivalGoods.objects.get(id=int(local_id))
-                arrival_Goods_obj.item_name = request.POST[item_name_list[i]]
-                arrival_Goods_obj.former_name = request.POST[former_name_list[i]]
-                arrival_Goods_obj.initial_qty = request.POST[qty_list[i]]
-                arrival_Goods_obj.qty = request.POST[qty_list[i]]
-                arrival_Goods_obj.advance = request.POST[advance_amount_list[i]]
-                arrival_Goods_obj.weight = request.POST[weight_list[i]]
-                arrival_Goods_obj.remarks = request.POST[remarks_list[i]]
-
-                arrival_Goods_obj.save()
-                print(f"Update Arrival Goods item  = {arrival_Goods_obj.id}")
-
-            else:
-
-                arrival_Goods_obj = ArrivalGoods(
-                    item_name=request.POST[item_name_list[i]],
-                    shop=shop_obj,
-                    arrival_entry=arrival,
-                    former_name=request.POST[former_name_list[i]],
-                    advance=request.POST[advance_amount_list[i]],
-                    initial_qty=request.POST[qty_list[i]],
-                    qty=request.POST[qty_list[i]],
-                    weight=request.POST[weight_list[i]],
-                    remarks=request.POST[remarks_list[i]],
-                )
-
-                arrival_Goods_obj.save()
-                print(f"New Arrival Goods item  = {arrival_Goods_obj.id}")
-    return render(request, 'index.html')
 
 
 @api_view(('GET',))
@@ -1152,88 +915,8 @@ def grouping_sales_bill_entry(sales_response_list: list):
     return response
 
 
-def generate_patti_pdf_bill(request):
-    if request.user.is_authenticated:
-        shop_detail_object = Shop.objects.get(shop_owner=request.user.id)
-
-        patti_entry_obj = PattiEntry.objects.get(
-            id=request.POST['patti_bill_id'])
-        patti_entry_obj.lorry_no = request.POST['patti_lorry_number']
-        patti_entry_obj.shop = shop_detail_object
-        patti_entry_obj.date = getDate_from_string(
-            request.POST['patti_entry_date'])
-        patti_entry_obj.advance = request.POST['advance_amount']
-        patti_entry_obj.farmer_name = request.POST['patti_farmer_name']
-        patti_entry_obj.total_weight = request.POST['total_weight']
-        patti_entry_obj.hamali = request.POST['hamali']
-        patti_entry_obj.net_amount = request.POST['net_amount']
-        patti_entry_obj.Empty_data = False
-
-        patti_entry_obj.save()
-        print(f"New Patti entry item  = {patti_entry_obj.id}")
-
-        arrival_detail_object = ArrivalEntry.objects.get(
-            shop=shop_detail_object,
-            lorry_no=request.POST['patti_lorry_number'],
-            date=getDate_from_string(request.POST['patti_entry_date']))
-
-        arrival_good_object = ArrivalGoods.objects.get(
-            shop=shop_detail_object,
-            arrival_entry=arrival_detail_object,
-            former_name=request.POST['patti_farmer_name'],
-
-        )
-        arrival_good_object.patti_status = True
-        arrival_good_object.save()
-
-        add_patti_item_list(request, list(request.POST), patti_entry_obj)
-        return Report.patti_report_view(request)
-        # return patti_list(request)
-    return render(request, 'index.html')
 
 
-def add_patti_item_list(request, request_list, patti_entry_obj):
-    if request.user.is_authenticated:
-        item_name_list = []
-        lot_number_list = []
-        weight_list = []
-        rate_list = []
-        amount_list = []
-
-        for i in request_list:
-            item_name_list_regrex = re.search("^.*_item_name$", i)
-            if item_name_list_regrex:
-                item_name_list.append(i)
-
-            lot_number_list_regrex = re.search("^.*_lot_number$", i)
-            if lot_number_list_regrex:
-                lot_number_list.append(i)
-
-            weight_list_regrex = re.search("^.*_weight$", i)
-            if weight_list_regrex:
-                weight_list.append(i)
-
-            rate_list_regrex = re.search("^.*_rate$", i)
-            if rate_list_regrex:
-                rate_list.append(i)
-
-            amount_list_regrex = re.search("^.*_amount$", i)
-            if amount_list_regrex:
-                amount_list.append(i)
-
-        for i in range(0, len(item_name_list)):
-            patti_obj = PattiEntryList(
-                item=request.POST[item_name_list[i]],
-                lot_no=request.POST[lot_number_list[i]],
-                weight=request.POST[weight_list[i]],
-                rate=request.POST[rate_list[i]],
-                amount=request.POST[amount_list[i]],
-                patti=patti_entry_obj
-            )
-
-            patti_obj.save()
-
-    return render(request, 'index.html')
 
 
 @csrf_protect
@@ -1262,22 +945,7 @@ def edit_sales_bill_entry(request, sales_id):
     return render(request, 'index.html')
 
 
-@csrf_protect
-def edit_patti_entry(request, patti_id):
-    if request.user.is_authenticated:
-        patti_bill_detail = PattiEntry.objects.get(pk=patti_id)
-        today = patti_bill_detail.date
 
-        patti_entry_obj = PattiEntryList.objects.filter(
-            patti=patti_bill_detail)
-
-        return render(request, 'Entry/Patti/modify_patti_entry.html',
-                      {'patti_bill_detail': patti_bill_detail,
-                       "today": today,
-                       "patti_entry_obj": patti_entry_obj,
-                       "NEW": False}
-                      )
-    return render(request, 'index.html')
 
 
 @api_view(['GET'])
